Suppress format string warnings
tool/uriparse.c:*:*: warning: field precision specifier ‘.*’ expects argument of type ‘int’, but argument 2 has type ‘long int’ [-Wformat]

+++ tool/uriparse.c
@@ -43,7 +43,7 @@
 #include <netinet/in.h>
 
 
-#define RANGE(x) ((x).afterLast-(x).first),((x).first)
+#define RANGE(x)  (int)((x).afterLast-(x).first), ((x).first)
 
 
 void usage() {
